[refactor] file structure
