+++ src/ui/components/Controls/Preview.svelte
