+++ src/ui/components/Layout/PreviewCard.svelte
