+++ src/ui/index.html
@@ -8,6 +8,6 @@
 	</head>
 	<body>
 		<div id="app"></div>
-		<script type="module" src="/src/ui/main.ts"></script>
+		<script type="module" src="./main.ts"></script>
 	</body>
 </html>
